Refactor everything to separate module
